ci: add nix env and starter gha workflows
Creates an initial skeleton for CI runs, based on the standard JS
tooling we support in other repos. The changes are unobjectionable:
containerfile for building images, nix env with justfile for common
local commands.

Also modified the app code slightly to accept an optional database CA
cert, which is necessary for the managed db services when leveraging
node-postgres; see related discussion in [0].

Overhauls the README based on all the above.

[0] https://github.com/penumbra-zone/dex-explorer/issues/55